+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 * Add the `space` as a hardcoded 4px-wide whitespace character
 * Improve getting a slot's coordinates
 * Make ProgressWidget register itself correctly
+* Allow adding "overlays" to `BaseSlot` instances